requirements.txt
disnake==2.9.*
python-dotenv==1.0.0
dataclasses-json